requirements.txt
# --- Núcleo científico / ML
torch>=2.2           # PyTorch: ya soporta Python 3.12
torchvision>=0.17
timm>=0.9.12         # Model zoo (EfficientNet)
xgboost>=2.0.3       # GPU / CPU compatible
scikit-learn>=1.5
imbalanced-learn>=0.12
scipy>=1.12
numpy>=1.26
pandas>=2.2
joblib>=1.3
packaging>=24.0

# --- Visualización & notebooks (EDA)
matplotlib>=3.9
seaborn>=0.13
tqdm>=4.66           # barra de progreso, opcional
ipython>=8.22

# --- Procesado de imágenes
pillow>=10.3

# --- API & serving
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
pydantic>=2.6
requests>=2.31       # para pruebas de la API

# --- Miscelánea / utilidades
watchfiles>=0.21     # recarga automática del servidor (uvicorn[standard] ya lo incluye, pero lo listamos explícito)
